Slette ubrukte og deprekerte controllere
Og oppdatere tester som ikke var oppdaterte
